Part of WOR-169: remove backward-compat shims, extract create_pr to watcher_subprocess
- Remove 6 noqa:F401 backward-compat re-export lines from watcher.py imports
- Remove 13 delegation shim methods (_create_worktree, _launch_worker, etc.)
  and update all call-sites to call watcher_worktrees/subprocess functions directly
- Remove 4 service shims (_ensure_ollama_running, _ensure_litellm_running,
  _stop_litellm_proxy, _litellm_proc property); call self._services.* directly
- Move _create_pr (88 LOC) to watcher_subprocess.create_pr() free function
- watcher.py: 844 → 656 LOC (Advisory tier)
- Update all test patches from patch.object(w, "_shim") to module-level patch targets
